fix: fix a bug in diagnostic messages
There was a bug where diagnostics messages could be sent by and to some
users but not to others.
The problem is with the MarkdownV2 parsing mode of telegram, just like
with the previous bug where some users and admins could send messages
whereas other cannot.
I solved it by stopping the user MarkdownV2 in diagnostic messages, and
instead, use the good old entities fields.